+++ schema.ts
@@ -2103,6 +2103,52 @@ export interface GqlSorSwapRouteHop {
 
 export type GqlSorSwapType = 'EXACT_IN' | 'EXACT_OUT';
 
+export interface GqlStakedSonicData {
+    __typename?: 'GqlStakedSonicData';
+    /** A list of all the delegated validators. */
+    delegatedValidators: Array<GqlStakedSonicDelegatedValidator>;
+    /** Current exchange rate for stS -> S */
+    exchangeRate: Scalars['String'];
+    /** The current rebasing APR for stS. */
+    stakingApr: Scalars['String'];
+    /** Total amount of S in custody of stS. Delegated S plus pool S. */
+    totalAssets: Scalars['AmountHumanReadable'];
+    /** Total amount of S elegated to validators. */
+    totalAssetsDelegated: Scalars['AmountHumanReadable'];
+    /** Total amount of S in the pool to be delegated. */
+    totalAssetsPool: Scalars['AmountHumanReadable'];
+}
+
+export interface GqlStakedSonicDelegatedValidator {
+    __typename?: 'GqlStakedSonicDelegatedValidator';
+    /** The amount of S that has been delegated to this validator. */
+    assetsDelegated: Scalars['AmountHumanReadable'];
+    /** The id of the validator. */
+    validatorId: Scalars['String'];
+}
+
+export interface GqlStakedSonicSnapshot {
+    __typename?: 'GqlStakedSonicSnapshot';
+    /** Current exchange rate for stS -> S */
+    exchangeRate: Scalars['String'];
+    id: Scalars['ID'];
+    /** The timestamp of the snapshot. Timestamp is end of day midnight. */
+    timestamp: Scalars['Int'];
+    /** Total amount of S in custody of stS. Delegated S plus pool S. */
+    totalAssets: Scalars['AmountHumanReadable'];
+    /** Total amount of S delegated to validators. */
+    totalAssetsDelegated: Scalars['AmountHumanReadable'];
+    /** Total amount of S in the pool. */
+    totalAssetsPool: Scalars['AmountHumanReadable'];
+}
+
+export type GqlStakedSonicSnapshotDataRange =
+    | 'ALL_TIME'
+    | 'NINETY_DAYS'
+    | 'ONE_HUNDRED_EIGHTY_DAYS'
+    | 'ONE_YEAR'
+    | 'THIRTY_DAYS';
+
 /** Inputs for the call data to create the swap transaction. If this input is given, call data is added to the response. */
 export interface GqlSwapCallDataInput {
     /** How long the swap should be valid, provide a timestamp. "999999999999999999" for infinite. Default: infinite */
@@ -2404,6 +2450,7 @@ export interface Mutation {
     poolReloadStakingForAllPools: Scalars['String'];
     poolSyncAllCowSnapshots: Array<GqlPoolMutationResult>;
     poolSyncAllPoolsFromSubgraph: Array<Scalars['String']>;
+    poolSyncFxQuoteTokens: Array<GqlPoolMutationResult>;
     poolUpdateLifetimeValuesForAllPools: Scalars['String'];
     poolUpdateLiquidityValuesForAllPools: Scalars['String'];
     protocolCacheMetrics: Scalars['String'];
@@ -2448,6 +2495,10 @@ export interface MutationPoolReloadStakingForAllPoolsArgs {
 }
 
 export interface MutationPoolSyncAllCowSnapshotsArgs {
+    chains: Array<GqlChain>;
+}
+
+export interface MutationPoolSyncFxQuoteTokensArgs {
     chains: Array<GqlChain>;
 }
 
@@ -2547,6 +2598,10 @@ export interface Query {
     sorGetSwapPaths: GqlSorGetSwapPaths;
     /** Get swap quote from the SOR, queries both the old and new SOR */
     sorGetSwaps: GqlSorGetSwapsResponse;
+    /** Get the staking data and status for stS */
+    stsGetGqlStakedSonicData: GqlStakedSonicData;
+    /** Get snapshots for sftmx staking for a specific range */
+    stsGetStakedSonicSnapshots: Array<GqlStakedSonicSnapshot>;
     /**
      * Returns the candlestick chart data for a token for a given range.
      * @deprecated Use tokenGetHistoricalPrices instead
@@ -2722,6 +2777,10 @@ export interface QuerySorGetSwapsArgs {
     swapType: GqlSorSwapType;
     tokenIn: Scalars['String'];
     tokenOut: Scalars['String'];
+}
+
+export interface QueryStsGetStakedSonicSnapshotsArgs {
+    range: GqlStakedSonicSnapshotDataRange;
 }
 
 export interface QueryTokenGetCandlestickChartDataArgs {
@@ -3083,6 +3142,10 @@ export type ResolversTypes = ResolversObject<{
     GqlSorSwapRoute: ResolverTypeWrapper<GqlSorSwapRoute>;
     GqlSorSwapRouteHop: ResolverTypeWrapper<GqlSorSwapRouteHop>;
     GqlSorSwapType: GqlSorSwapType;
+    GqlStakedSonicData: ResolverTypeWrapper<GqlStakedSonicData>;
+    GqlStakedSonicDelegatedValidator: ResolverTypeWrapper<GqlStakedSonicDelegatedValidator>;
+    GqlStakedSonicSnapshot: ResolverTypeWrapper<GqlStakedSonicSnapshot>;
+    GqlStakedSonicSnapshotDataRange: GqlStakedSonicSnapshotDataRange;
     GqlSwapCallDataInput: GqlSwapCallDataInput;
     GqlToken: ResolverTypeWrapper<GqlToken>;
     GqlTokenAmountHumanReadable: GqlTokenAmountHumanReadable;
@@ -3259,6 +3322,9 @@ export type ResolversParentTypes = ResolversObject<{
     GqlSorSwapOptionsInput: GqlSorSwapOptionsInput;
     GqlSorSwapRoute: GqlSorSwapRoute;
     GqlSorSwapRouteHop: GqlSorSwapRouteHop;
+    GqlStakedSonicData: GqlStakedSonicData;
+    GqlStakedSonicDelegatedValidator: GqlStakedSonicDelegatedValidator;
+    GqlStakedSonicSnapshot: GqlStakedSonicSnapshot;
     GqlSwapCallDataInput: GqlSwapCallDataInput;
     GqlToken: GqlToken;
     GqlTokenAmountHumanReadable: GqlTokenAmountHumanReadable;
@@ -4963,6 +5029,41 @@ export type GqlSorSwapRouteHopResolvers<
     __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
 }>;
 
+export type GqlStakedSonicDataResolvers<
+    ContextType = ResolverContext,
+    ParentType extends ResolversParentTypes['GqlStakedSonicData'] = ResolversParentTypes['GqlStakedSonicData'],
+> = ResolversObject<{
+    delegatedValidators?: Resolver<Array<ResolversTypes['GqlStakedSonicDelegatedValidator']>, ParentType, ContextType>;
+    exchangeRate?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    stakingApr?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    totalAssets?: Resolver<ResolversTypes['AmountHumanReadable'], ParentType, ContextType>;
+    totalAssetsDelegated?: Resolver<ResolversTypes['AmountHumanReadable'], ParentType, ContextType>;
+    totalAssetsPool?: Resolver<ResolversTypes['AmountHumanReadable'], ParentType, ContextType>;
+    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+}>;
+
+export type GqlStakedSonicDelegatedValidatorResolvers<
+    ContextType = ResolverContext,
+    ParentType extends ResolversParentTypes['GqlStakedSonicDelegatedValidator'] = ResolversParentTypes['GqlStakedSonicDelegatedValidator'],
+> = ResolversObject<{
+    assetsDelegated?: Resolver<ResolversTypes['AmountHumanReadable'], ParentType, ContextType>;
+    validatorId?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+}>;
+
+export type GqlStakedSonicSnapshotResolvers<
+    ContextType = ResolverContext,
+    ParentType extends ResolversParentTypes['GqlStakedSonicSnapshot'] = ResolversParentTypes['GqlStakedSonicSnapshot'],
+> = ResolversObject<{
+    exchangeRate?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
+    timestamp?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
+    totalAssets?: Resolver<ResolversTypes['AmountHumanReadable'], ParentType, ContextType>;
+    totalAssetsDelegated?: Resolver<ResolversTypes['AmountHumanReadable'], ParentType, ContextType>;
+    totalAssetsPool?: Resolver<ResolversTypes['AmountHumanReadable'], ParentType, ContextType>;
+    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+}>;
+
 export type GqlTokenResolvers<
     ContextType = ResolverContext,
     ParentType extends ResolversParentTypes['GqlToken'] = ResolversParentTypes['GqlToken'],
@@ -5240,6 +5341,12 @@ export type MutationResolvers<
         RequireFields<MutationPoolSyncAllCowSnapshotsArgs, 'chains'>
     >;
     poolSyncAllPoolsFromSubgraph?: Resolver<Array<ResolversTypes['String']>, ParentType, ContextType>;
+    poolSyncFxQuoteTokens?: Resolver<
+        Array<ResolversTypes['GqlPoolMutationResult']>,
+        ParentType,
+        ContextType,
+        RequireFields<MutationPoolSyncFxQuoteTokensArgs, 'chains'>
+    >;
     poolUpdateLifetimeValuesForAllPools?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
     poolUpdateLiquidityValuesForAllPools?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
     protocolCacheMetrics?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
@@ -5439,6 +5546,13 @@ export type QueryResolvers<
         ParentType,
         ContextType,
         RequireFields<QuerySorGetSwapsArgs, 'swapAmount' | 'swapOptions' | 'swapType' | 'tokenIn' | 'tokenOut'>
+    >;
+    stsGetGqlStakedSonicData?: Resolver<ResolversTypes['GqlStakedSonicData'], ParentType, ContextType>;
+    stsGetStakedSonicSnapshots?: Resolver<
+        Array<ResolversTypes['GqlStakedSonicSnapshot']>,
+        ParentType,
+        ContextType,
+        RequireFields<QueryStsGetStakedSonicSnapshotsArgs, 'range'>
     >;
     tokenGetCandlestickChartData?: Resolver<
         Array<ResolversTypes['GqlTokenCandlestickChartDataItem']>,
@@ -5677,6 +5791,9 @@ export type Resolvers<ContextType = ResolverContext> = ResolversObject<{
     GqlSorSwap?: GqlSorSwapResolvers<ContextType>;
     GqlSorSwapRoute?: GqlSorSwapRouteResolvers<ContextType>;
     GqlSorSwapRouteHop?: GqlSorSwapRouteHopResolvers<ContextType>;
+    GqlStakedSonicData?: GqlStakedSonicDataResolvers<ContextType>;
+    GqlStakedSonicDelegatedValidator?: GqlStakedSonicDelegatedValidatorResolvers<ContextType>;
+    GqlStakedSonicSnapshot?: GqlStakedSonicSnapshotResolvers<ContextType>;
     GqlToken?: GqlTokenResolvers<ContextType>;
     GqlTokenCandlestickChartDataItem?: GqlTokenCandlestickChartDataItemResolvers<ContextType>;
     GqlTokenData?: GqlTokenDataResolvers<ContextType>;